Create Tic-Tac-Toe-Code
The code for this program uses a 2D array to hold a char value of ' ', 'X', or 'O'. The "board" of the game is shown at the beginning and after a change is made to the "board". This program uses similar code for player 1 and player 2, but with slight modification. The skills required to create this program include knowledge of functions (pass by value, pass by reference, and prototypes), 2D arrays, if statements, loops, and the bool data type.